MVVM : Namespace rename
