+++ src/VideoManagement/VideoManagement.DataAccess/SQLite.Application/ApplicationDBContext.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
-using VideoManagement.Models;
+using VideoManagement.Models.Tables;
 
 namespace VideoManagement.DataAccess.SQLite.Application
 {
