+++ src/VideoManagement/VideoManagement.DataAccess/SQLite.Application/Configuration/RecentPathsMap.cs
@@ -3,7 +3,7 @@ using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using System;
 using System.Collections.Generic;
 using System.Text;
-using VideoManagement.Models;
+using VideoManagement.Models.Tables;
 
 namespace VideoManagement.DataAccess.SQLite.Application.Configuration
 {
