+++ src/VideoManagement/VideoManagement.Models/Tables/Artist.cs
@@ -2,9 +2,9 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace VideoManagement.Models
+namespace VideoManagement.Models.Tables
 {
-    public class Category : BaseModel
+    public class Artist : BaseModel
     {
         public string Name { get; set; }
     }
