+++ src/VideoManagement/VideoManagement.UI.WPF/ViewModels/MainWindowModel.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using VideoManagement.Models;
+using VideoManagement.Models.Tables;
 
 namespace VideoManagement.UI.WPF.ViewModels
 {
